requirements.txt
sounddevice
vosk
ttkbootstrap
pygame
pyttsx3
edge-tts
screen-brightness-control
Pillow
PyGetWindow
PyAutoGUI
opencv-python
numpy
pyaudio
moviepy
wikipedia
beautifulsoup4
requests
speedtest-cli
googletrans==4.0.0-rc1
